Refactoring 'processes' package
